Add PaneLayout::view_with closure-based pane rendering
Inherent method on PaneLayout that draws chrome and invokes a per-pane
render closure with a chrome-inset RenderContext whose chrome_owned is
set to true. Replaces the consumer-side three-step dance of state.layout()
+ chrome render + manual Margin{1,1} inner-rect computation.

Component::view trait impl gets an explanatory docstring noting it
renders chrome only — generic-Component code should call view_with on
the concrete type for embedded children. Its body now delegates to a
private render_chrome_only helper sharing the chrome-drawing path with
view_with, which deduplicates pane rendering and keeps mod.rs under the
1000-line project limit.

view_with's docstring warns against constructing fresh RenderContext::new
inside the closure (bypasses chrome_owned and re-introduces double-render).

Tracks D2.
